I added my first project in python here
note: 15% of the click game inside of the project was made using ai becouse I had truble with space. (I will get better I did this as a part of school activity I had to finish it in time,everything else in the project was all made by me)
advice: use ai for creating a template to experment on to understand the code better basically salf teach yourself.